Bundle .js files
