+++ public/js/component/carousel.js
@@ -1,5 +1,7 @@
+import { find, findOne } from "../util"
+
 // append a simple image-carousel to the wrapper
-function appendImageCarousel(wrapper, id, items, pageSize = 1) {
+export function appendImageCarousel(wrapper, id, items, pageSize = 1) {
   const carouselItemHtmlString = items.reduce((str, { href, src }) => str + `
     <div class="carousel-item w-${Math.floor(100 / pageSize)}">
       <a href="${href}">
@@ -21,7 +23,7 @@ function appendImageCarousel(wrapper, id, items, pageSize = 1) {
 }
 
 // initialize a carousel
-function initCarousel(carousel) {
+export function initCarousel(carousel) {
   const carouselId = carousel.getAttribute('id')
   const items = find(`#${carouselId} .carousel-item`)
   const carouselInfo = {
@@ -48,7 +50,7 @@ function initCarousel(carousel) {
 }
 
 // show next carousel item (also support reverse order)
-function showNextCarouselItem(carouselInfo, isReversed = false) {
+export function showNextCarouselItem(carouselInfo, isReversed = false) {
   const { pageSize, items, nItems, isOnTransition } = carouselInfo
   let { index } = carouselInfo
