+++ public/js/component/index.js
@@ -1,0 +1,4 @@
+export * from './carousel'
+export * from './hover_popup'
+export * from './item_card'
+export * from './recent'
