+++ public/js/util/index.js
@@ -1,0 +1,4 @@
+export * from './mypromise'
+export * from './myfetch'
+export * from './myqueryselector'
+export * from './local_storage'
