+++ public/js/util/myfetch.js
@@ -1,4 +1,6 @@
-function myFetch(url, init = {}) {
+import MyPromise from './mypromise'
+
+export function myFetch(url, init = {}) {
   return new MyPromise((resolve, reject) => {
     try {
       const {
@@ -44,4 +46,24 @@ function myFetch(url, init = {}) {
       reject(error)
     }
   })
+}
+
+// fetch item or item-list from the server
+export function fetchItems(route, data = {}) {
+  // make query-string
+  if (Object.keys(data).length) {
+    route += '?' + objectToQueryString(data)
+  }
+
+  return myFetch(`/items${route}`)
+    .then(response => JSON.parse(response))
+}
+
+// convert object to query-string
+// only support simple object: no array or object as a property
+function objectToQueryString(obj) {
+  return Object
+    .keys(obj)
+    .map(key => `${encodeURIComponent(key)}=${encodeURIComponent(obj[key])}`)
+    .join('&')
 }
