+++ views/index.html
@@ -104,17 +104,7 @@
     <footer class="p-5 mt-5">footer</footer>
   </div>
 
-  <script src="js/util.js"></script>
-  <script src="js/mypromise.js"></script>
-  <script src="js/myfetch.js"></script>
-  <script src="js/myqueryselector.js"></script>
-  <script src="js/local_storage.js"></script>
-  <script src="js/hover_popup.js"></script>
-
-  <script src="js/component/recent.js"></script>
-  <script src="js/component/carousel.js"></script>
-  <script src="js/component/item_card.js"></script>
-  <script src="js/index.js"></script>
+  <script src="js/dist/script.bundle.js"></script>
 </body>
 
 </html>
